Move editing test into thread.spec.ts
